+++ Example/Visualization/ScatterPlotMatrix/main.cpp
@@ -8,8 +8,8 @@
 #include <kvs/glut/Application>
 #include <kvs/glut/Screen>
 #include <kvs/TableObject>
-#include <kvs/ScatterPlotRenderer>
-#include <kvs/Axis2D>
+#include <kvs/ScatterPlotMatrixRenderer>
+#include <kvs/Axis2DMatrix>
 
 
 /*===========================================================================*/
@@ -24,17 +24,19 @@ int main( int argc, char** argv )
 {
     kvs::glut::Application app( argc, argv );
     kvs::glut::Screen screen( &app );
-    screen.setTitle( "kvs::ScatterPlotRenderer" );
+    screen.setTitle( "kvs::ScatterPlotMatrixRenderer" );
 
-    const size_t nsamples = 100;
+    const size_t nsamples = 50;
     const float min_value = 0.0f;
     const float max_value = 1.0f;
     const auto random_x = kvs::ValueArray<float>::Random( nsamples, min_value, max_value );
     const auto random_y = kvs::ValueArray<float>::Random( nsamples, min_value, max_value );
+    const auto random_z = kvs::ValueArray<float>::Random( nsamples, min_value, max_value );
 
     auto* object = new kvs::TableObject();
-    object->addColumn( kvs::AnyValueArray( random_x ), "Random X" );
-    object->addColumn( kvs::AnyValueArray( random_y ), "Random Y" );
+    object->addColumn( kvs::AnyValueArray( random_x ), "X value" );
+    object->addColumn( kvs::AnyValueArray( random_y ), "Y value" );
+    object->addColumn( kvs::AnyValueArray( random_z ), "Z value" );
 
     object->setMinValue( 0, min_value );
     object->setMaxValue( 0, max_value );
@@ -42,23 +44,24 @@ int main( int argc, char** argv )
     object->setMinValue( 1, min_value );
     object->setMaxValue( 1, max_value );
 
+    object->setMinValue( 2, min_value );
+    object->setMaxValue( 2, max_value );
+
     const int top_margin = 35;
     const int left_margin = 65;
     const int bottom_margin = 50;
 
-    auto* renderer = new kvs::ScatterPlotRenderer();
+    auto* renderer = new kvs::ScatterPlotMatrixRenderer();
     renderer->setTopMargin( top_margin );
     renderer->setLeftMargin( left_margin );
     renderer->setBottomMargin( bottom_margin );
-    renderer->setPointSize( 5.0f );
+    renderer->setPointSize( 4.0f );
 
-    auto* axis = new kvs::Axis2D();
+    auto* axis = new kvs::Axis2DMatrix();
     axis->setTitle( "Title" );
     axis->setTopMargin( top_margin );
     axis->setLeftMargin( left_margin );
     axis->setBottomMargin( bottom_margin );
-    axis->xAxis().setLabel( "X value" );
-    axis->yAxis().setLabel( "Y value" );
 
     screen.registerObject( object, axis );
     screen.registerObject( object, renderer );